Package tablewriter: Use a pointer receiver for Table's methods.
This allows to use "defer" to print the table, for example:

  w := NewWriter(os.Stdout)
  defer w.Render()
  for _, data := range ch {
    w.Append(data)
  }
  return

Previously, the "defer w.Render()" line would create a copy of "w"
which, at the time, contained an empty "lines" slice, so that the table
printed was always empty.

Last but not least: passing the receiver by reference should be more
efficient for big structs, such as Table.