Add support for multi-line headers and footers.
Also add a new option "reflow during auto wrap"
(`SetReflowDuringAutoWrap()`), which when enabled merges all the lines
in a multi-line cell together before to wrapping, and when disabled
causes paragraphs in a multi-line cell to be cleanly separated.

The default is true to preserve the previous behavior in existing
code.